Redesign the Operator
Now, only FunctionClass and Expr can be argument of Operator.
