+++ sympy/core/function.py
@@ -266,6 +266,71 @@ class FunctionClass(ManagedProperties):
     def __repr__(cls):
         return cls.__name__
 
+    def __pos__(cls):
+        from sympy.operator import Operator
+        return Operator(cls)
+
+    def __neg__(cls):
+        from sympy.operator import Operator
+        return -Operator(cls)
+
+    def __add__(cls, other):
+        from sympy.operator import OperatorAdd
+        return OperatorAdd(cls, other)
+
+    def __radd__(cls, other):
+        from sympy.operator import OperatorAdd
+        return OperatorAdd(other, cls)
+
+    def __sub__(cls, other):
+        from sympy.operator import OperatorAdd
+        return OperatorAdd(cls, -other)
+
+    def __rsub__(cls, other):
+        from sympy.operator import OperatorAdd
+        return OperatorAdd(-cls, other)
+
+    def __mul__(cls, other):
+        from sympy.operator import OperatorMul
+        return OperatorMul(cls, other)
+
+    def __rmul__(cls, other):
+        from sympy.operator import OperatorMul
+        return OperatorMul(other, cls)
+
+    def __pow__(self, other):
+        from sympy.operator import OperatorPow
+        return OperatorPow(self, other)
+
+    def __rpow__(self, other):
+        from sympy.operator import OperatorPow
+        return OperatorPow(other, self)
+
+    def __div__(self, other):
+        from sympy.core.singleton import S
+        from sympy.operator import OperatorMul, OperatorPow
+        return OperatorMul(self, OperatorPow(other, S.NegativeOne))
+
+    def __rdiv__(self, other):
+        from sympy.core.singleton import S
+        from sympy.operator import OperatorMul, OperatorPow
+        return OperatorMul(other, OperatorPow(self, S.NegativeOne))
+
+    def __matmul__(self, other):
+        from sympy.operator import CompositeOperator
+        return CompositeOperator(self, other)
+
+    def __rmatmul__(self, other):
+        from sympy.operator import CompositeOperator
+        return CompositeOperator(other, self)
+
+    def deriv(self, *argidxs, **kwargs):
+        from sympy.operator import DerivatedOperator
+        return DerivatedOperator(self, *argidxs, **kwargs)
+
+    __truediv__ = __div__
+    __rtruediv__ = __rdiv__
+
 
 class Application(with_metaclass(FunctionClass, Basic)):
     """
