+++ sympy/operator/composite.py
@@ -1,12 +1,7 @@
 """Composite operator."""
 
 from .operator import OperatorExpr, Operator
-from sympy.core.add import Add
-from sympy.core.compatibility import iterable
 from sympy.core.containers import Tuple
-from sympy.core.mul import Mul
-from sympy.core.power import Pow
-from sympy.core.singleton import S
 
 class CompositeOperator(OperatorExpr):
     """
@@ -29,10 +24,10 @@ class CompositeOperator(OperatorExpr):
 
     Parameters
     ==========
-    f : instance of Basic or BasicMeta
+    f : instance of Expr or FunctionClass
             Operator which will be evaluated.
 
-    gs : tuple of Basic or BasicMeta
+    gs : tuple of Expr or FunctionClass
             Operators which will be passed as arguments of f.
     """
     def __new__(cls, f, *gs, **kwargs):
