+++ sympy/operator/operatorop.py
@@ -226,7 +226,7 @@ class OperatorAdd(OperatorExpr, Add):
             return newseq, [], None
 
     def __new__(cls, *args, **kwargs):
-        new_args = cls._process_basic_basicmeta(*args)
+        new_args = cls._process_operator(*args)
         obj = super(cls, OperatorAdd).__new__(cls, *new_args, **kwargs)
         if not isinstance(obj, OperatorExpr):
             return Operator(obj)
@@ -258,7 +258,7 @@ class OperatorMul(OperatorExpr, Mul):
     identity = S.One
 
     def __new__(cls, *args, **kwargs):
-        new_args = cls._process_basic_basicmeta(*args)
+        new_args = cls._process_operator(*args)
         obj = super(cls, OperatorMul).__new__(cls, *new_args, **kwargs)
         if not isinstance(obj, OperatorExpr):
             return Operator(obj)
@@ -774,7 +774,7 @@ class OperatorPow(OperatorExpr, Pow):
     sin(x)**x
     """
     def __new__(cls, b, e, **kwargs):
-        new_b, new_e = cls._process_basic_basicmeta(b, e)
+        new_b, new_e = cls._process_operator(b, e)
         obj = super(cls, OperatorPow).__new__(cls, new_b, new_e, **kwargs)
         if not isinstance(obj, OperatorExpr):
             return Operator(obj)
