+++ sympy/operator/tests/test_operatorop.py
@@ -1,4 +1,4 @@
-from sympy import Add, cos, Function, sin, symbols
+from sympy import cos, Function, sin, symbols
 from sympy.abc import x,y,z
 from sympy.operator import Operator
 
@@ -10,8 +10,6 @@ def test_OperatorAdd():
     assert (sin-cos)(x) == sin(x)-cos(x)
     assert (sin+1-sin)(x) == 1
 
-    assert (sin+Add)(x,y,z) == sin(x)+x+y+z
-    assert (Operator(sin, (1,)) + Operator(Add, (0,2)))(x,y,z) == sin(y) + x+z
 
 def test_OperatorMul():
     assert (1*f)(x) == f(x)
